requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
aiosqlite==0.21.0
pyyaml==6.0.2
aiohttp
psutil==7.0.0
huggingface_hub==0.34.4
